requirements.txt
reportlab>=4.0.0
openai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
